Station Officer side activated.
-rent history can see
-can change availability
