+++ app/src/main/java/com/example/padelgo/stationOfficer/AdminDash.java
@@ -1,4 +1,4 @@
-package com.example.padelgo.admin;
+package com.example.padelgo.stationOfficer;
 
 import android.annotation.SuppressLint;
 import android.content.Intent;
