+++ app/src/main/java/com/example/padelgo/stationOfficer/AdminLocation.java
@@ -1,4 +1,4 @@
-package com.example.padelgo.admin;
+package com.example.padelgo.stationOfficer;
 
 import android.content.Intent;
 import android.os.Bundle;
