+++ app/src/main/res/layout/activity_admin_account_create.xml
@@ -8,7 +8,7 @@
     android:background="@drawable/bg"
     android:gravity="center"
     android:orientation="vertical"
-    tools:context=".admin.AdminAccountCreate">
+    tools:context=".stationOfficer.AdminAccountCreate">
 
     <ImageView
         android:layout_width="200dp"
