+++ app/src/main/res/layout/activity_admin_dash.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/bg"
-    tools:context=".admin.AdminDash">
+    tools:context=".stationOfficer.AdminDash">
 
     <View
         android:id="@+id/top_view"
