+++ app/src/main/res/layout/activity_admin_profile.xml
@@ -8,7 +8,7 @@
     android:background="@drawable/bg"
     android:fillViewport="true"
     android:orientation="horizontal"
-    tools:context=".admin.AdminProfile">
+    tools:context=".stationOfficer.AdminProfile">
 
     <LinearLayout
         android:layout_width="match_parent"
