+++ app/src/main/res/layout/activity_admin_rents.xml
@@ -8,7 +8,7 @@
     android:background="@drawable/bg"
     android:fillViewport="true"
     android:orientation="horizontal"
-    tools:context=".admin.AdminRents">
+    tools:context=".stationOfficer.AdminRents">
 
     <LinearLayout
         android:layout_width="match_parent"
